new services created, service renaming, and listeners
Renamed services and listeners to maintain code clarity, added service that adds/deletes a guild from the bot's registry, created service to monitor members' connection time.

+++ src/main/java/com/wlmd/discord_bot/model/MemberActivityModel.java
@@ -54,12 +54,12 @@ public class MemberActivityModel {
 	private String sessionEnd;
 	
 	@Column(nullable = true)
-	private int totalTime;
+	private Long totalTime;
 	
 	public MemberActivityModel() {}
 
 	public MemberActivityModel(MemberModel member, Long memberActivityId, Long discordUserId, Long guildId, String nickName,
-			String lastSeen, String sessionStart, String sessionEnd, int totalTime) {
+			String lastSeen, String sessionStart, String sessionEnd, Long totalTime) {
 
 		this.member = member;
 		this.memberActivityId = memberActivityId;
@@ -76,7 +76,7 @@ public class MemberActivityModel {
 	    this.member = user;
 	    this.discordUserId = user.getDiscordUserId();
 	    this.guildId = user.getGuild().getGuildId();
-	    this.totalTime = 0;
+	    this.totalTime = 0L;
 	}
 
 	public Long getMemberActivityId() {
@@ -135,11 +135,11 @@ public class MemberActivityModel {
 		this.sessionEnd = sessionEnd;
 	}
 
-	public int getTotalTime() {
+	public Long getTotalTime() {
 		return totalTime;
 	}
 
-	public void setTotalTime(int totalTime) {
+	public void setTotalTime(Long totalTime) {
 		this.totalTime = totalTime;
 	}
